@@ -10,7 +10,7 @@ import {
   Status,
 } from 'starknet'
 
-// StarkNet network: Either goerli-alpha or mainnet-alpha
+// Starknet network: Either goerli-alpha or mainnet-alpha
 const network = 'goerli-alpha'
 
 /** Environment variables for a deployed and funded account to use for deploying contracts
