@@ -32,7 +32,7 @@ import (
 // --> privKey 2137244795266879235401249500471353867704187908407744160927664772020405449078 r 2898571078985034687500959842265381508927681132188252715370774777831313601543 s 1930849708769648077928186998643944706551011476358007177069185543644456022504 pubKey 1118148281956858477519852250235501663092798578871088714409528077622994994907
 //     privKey 3571531812827697194985986636869245829152430835021673171507607525908246940354 r 3242770073040892094735101607173275538752888766491356946211654602282309624331 s 2150742645846855766116236144967953798077492822890095121354692808525999221887 pubKey 2445157821578193538289426656074203099996547227497157254541771705133209838679
 
-func TestStarkNetKeyring_TestVector(t *testing.T) {
+func TestStarknetKeyring_TestVector(t *testing.T) {
 	var kr1 OCR2Key
 	bigKey, _ := new(big.Int).SetString("2137244795266879235401249500471353867704187908407744160927664772020405449078", 10)
 	feltKey := junotypes.BigToFelt(bigKey)
@@ -101,7 +101,7 @@ func TestStarkNetKeyring_TestVector(t *testing.T) {
 	assert.Equal(t, feltSExpected, s)
 }
 
-func TestStarkNetKeyring_Sign_Verify(t *testing.T) {
+func TestStarknetKeyring_Sign_Verify(t *testing.T) {
 	kr1, err := NewOCR2Key(cryptorand.Reader)
 	require.NoError(t, err)
 	kr2, err := NewOCR2Key(cryptorand.Reader)
@@ -163,7 +163,7 @@ func TestStarkNetKeyring_Sign_Verify(t *testing.T) {
 	})
 }
 
-func TestStarkNetKeyring_Marshal(t *testing.T) {
+func TestStarknetKeyring_Marshal(t *testing.T) {
 	kr1, err := NewOCR2Key(cryptorand.Reader)
 	require.NoError(t, err)
 	m, err := kr1.Marshal()
