@@ -16,6 +16,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/chainlink-starknet/ops"
+
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/keys"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/keys/mocks"
 	txmmock "github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/txm/mocks"
@@ -23,7 +24,7 @@ import (
 )
 
 func TestIntegration_Txm(t *testing.T) {
-	url := ops.SetupLocalStarkNetNode(t)
+	url := ops.SetupLocalStarknetNode(t)
 	rawLocalKeys := ops.TestKeys(t, 2) // generate 2 keys
 
 	// parse keys into expected format
@@ -88,7 +89,6 @@ func TestIntegration_Txm(t *testing.T) {
 
 	// start txm + checks
 	require.NoError(t, txm.Start(context.Background()))
-	require.NoError(t, txm.Healthy())
 	require.NoError(t, txm.Ready())
 
 	for k := range localKeys {
